Extract a test util func
